Refactoring #1 - complete rewrite of I2C implementation, moving classes between .api and .devices, removed jdkdio, pi4j and wiringpi providers.

+++ diozero-core/src/main/java/com/diozero/util/BitManipulation.java
@@ -63,7 +63,7 @@ public class BitManipulation {
 		return (byte)(1 << bit);
 	}
 
-	public static boolean isBitSet(byte value, byte bit) {
+	public static boolean isBitSet(byte value, int bit) {
 		return (value & getBitMask(bit)) != 0;
 	}
 }
